requirement.txt
PyQt6			6.8.0
PyQt6-Qt6		6.8.1
PyQt6-WebEngine		6.8.0
PyQt6-WebEngine-Qt6	6.8.1
PyQt6-sip		13.9.1
cachetools		5.5.1
certifi			2024.12.14
charset-normalizer	3.4.1
dateparser		1.2.0
deprecation		2.1.0
grpcio			1.70.0
idna			3.10
numpy			2.0.2
packaging		24.2
pandas			2.2.3
pip			24.3.1
plotly			5.24.1
protobuf		4.25.6
pybit			5.8.0
pycryptodome		3.21.0
python-dateutil		2.9.0.post0
pytz			2024.2
regex			2024.11.6
requests		2.32.3
setuptools		68.2.0
six			1.17.0
tenacity		9.0.0
tinkoff			0.1.1
tinkoff-investments	0.2.0b109	
tzdata			2024.2
tzlocal			5.2
urllib3			2.3.0
websocket-client	1.8.0
wheel			0.45.1

python			3.9.7